reworking of generate_report_files.R, updating of master RDS

+++ code/generate_report_files.R
@@ -1,17 +1,21 @@
-# Loads data and calls census API if necessary
-source(here("code", "census_data.R"))
-# Converts block group demographics to police district demographics
-source(here("code", "bg_to_policedist.R"))
-# Maintains and loads previously uploaded police district (master) shp files
-source(here("code", "update_policedist_shpfiles.R"))
-# Functions to create custom ggplot map objects
-source(here("code", "map_generator.R"))
-
-generate_analysis <- function(file, dist_name_var, geometry_var, city, ethnic_group = "Total", year = NULL) {
+generate_analysis <- function(
+    city, # city name (raleigh, charlotte, durham, etc.)
+    map_unit, # "city" or "district" 
+    dist_name = NULL, # if map_unit == "district", dist_name is district name
+    file = NULL, # if city is not in district_calculations, file path required
+    dist_name_var = NULL, # required with file, name of district name variable
+    geometry_var = NULL, # required with file, name of geometry variable
+    ethnic_group = "Total", # default is Total (no ethnic group, full population)
+    year = NULL # year of census data, (default is most current year available)
+    ) {
+  
   city_lower <- str_to_lower(city)
   
   # If current police district is not in master district file, update.
   if (!(city_lower %in% current_policedistricts$city)) { 
+    if (is.null(file) | is.null(dist_name_var) | is.null(geometry_var)) {
+      stop("Must provide file, district name variable, and geometry variable. City is not in current list.")
+    }
     # from update_policedist_shpfiles -- automatically updates environment variables
     append_shp(file, city_lower, dist_name_var, geometry_var)
   }
@@ -22,25 +26,30 @@ generate_analysis <- function(file, dist_name_var, geometry_var, city, ethnic_gr
     mutate(policedist_full_area = st_area(geometry)) |>
     st_make_valid(police_curr)
   
+  if (!is.null(dist_name)) {
+    police_curr <- police_curr |>
+      filter(DISTRICT == dist_name)
+  }
+  
   # Metadata
   num_dist <- nrow(police_curr)
-  name_dist <- police_curr |> pull(dist_name_var)
+  name_dist <- police_curr |> pull(DISTRICT)
   
   # ggplot #1, simple police district map. function from map_generator
-  police_dist_map <- police_district_map(police_curr, city_lower)
+  police_dist_map <- police_district_map(police_curr, city_lower, map_unit)
   
   # If user does not specify year, use currently-maintained current year
   if (is.null(year)) {
-    yr <- read_csv('data/census_data/census_data_metadata.csv') |> 
+    census_year <- read_csv('data/census_data/census_data_metadata.csv') |> 
       filter(status == "current") |> 
       pull(year)
   } else {
-    yr <- year
+    census_year <- year
   }
   
   # Load census data, calling API if current year is bg_to_policedist available
   # function from census_data
-  census_info <- read_census_data(yr)
+  census_info <- read_census_data(census_year)
   nc_bg_sf <- census_info$bgsf[[1]]
   acs_data_tbl <- census_info$acstbl[[1]]
   
@@ -54,9 +63,11 @@ generate_analysis <- function(file, dist_name_var, geometry_var, city, ethnic_gr
   
   # functions from bg_to_policedist -- geography recalculations
   city_intersection_sf <- bgtbl_to_bgsf(city_bg_tbl, police_curr)
+  police_dist_sf <- bgsf_to_poldistsf(city_intersection_sf)
   bg_overlap <- all_bg_overlapping_dist(city_intersection_sf, city_bg_tbl)
   
-  # Metadata, also used in map generation 
+  # Metadata
+  date_added <- Sys.Date()
   # Number of total touching census neighborhoods 
   total_bg <- nrow(bg_overlap |> group_by(GEOID) |> slice(1))
   # Number of census neighborhoods fully within a district
@@ -71,27 +82,28 @@ generate_analysis <- function(file, dist_name_var, geometry_var, city, ethnic_gr
     fully_included_bg)
   
   # ggplot #4, neighborhood-district specific populations. function from map_generator
-  dist_bg_pop_map <- resident_intersection_map(bg_overlap, police_curr, city_lower)
+  dist_bg_pop_map <- resident_intersection_map(bg_overlap, police_curr, city_lower, map_unit)
   
   # ggplot #5, police district populations
   dist_only_pop_map <- dist_population_map(
     police_dist_sf, 
     bg_overlap, 
     city, 
-    map_unit, 
-    perc_total, 
+    "city", 
     ethnic_group
   )
   
-  # Rbind to nested dataframe structure stored in memory:
-  district_calculations <- readRDS("data/district_calculations.rds")
+  # Join new objects together
   new_district_objects <- list(
     city_lower,
-    list(police_curr),
+    census_year,
+    date_added,
     num_dist,
-    name_dist,
+    list(name_dist),
+    list(police_curr),
     total_bg,
     fully_included_bg,
+    list(police_dist_sf),
     list(nc_bg_sf),
     list(police_dist_map),
     list(bg_pop_map),
@@ -99,8 +111,31 @@ generate_analysis <- function(file, dist_name_var, geometry_var, city, ethnic_gr
     list(dist_bg_pop_map),
     list(dist_only_pop_map)
   )
-  
-  joined_district_calculations <- rbind(district_calculations, new_district_objects)
-  saveRDS(joined_district_calculations, "data/district_calculations.rds")
 }
 
+append_analysis <- function(new_district_objects) {
+  master_analysis_path <- "data/district_calculations.rds"
+  if (file.exists(master_analysis_path)) {
+    district_calculations <- readRDS(master_analysis_path)
+    joined_district_calculations <- rbind(district_calculations, new_district_objects)
+    saveRDS(joined_district_calculations, "data/district_calculations.rds")
+  }
+  else {
+    names_master_analysis <- c(
+      "city_name", "census_year", "date_added", 
+      "district_num", "district_names",
+      "police_district_df", 
+      "total_bg_num", "fullyinc_bg_num",
+      "policedist_sf_df", "bg_sf_df", 
+      "police_dist_ggplot", 
+      "bg_population_ggplot", 
+      "dist_bg_areraintersection_ggplot", 
+      "dist_bg_numresident_ggplot", 
+      "dist_pop_map_ggplot"
+    )
+    
+    names(new_district_objects) <- names_master_analysis
+    new_obj_tibble <- as_tibble(new_district_objects)
+    saveRDS(new_obj_tibble, "data/district_calculations.rds")
+  }
+}
